refactor(core): Consolidate queries
