+++ src/main/java/org/sustain/controllers/OsmController.java
@@ -12,6 +12,7 @@ import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.sustain.OsmRequest;
 import org.sustain.SpatialOp;
+import org.sustain.db.queries.SpatialQueryUtil;
 import org.sustain.util.Constants;
 import org.sustain.db.mongodb.DBConnection;
 
@@ -19,7 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.LinkedBlockingQueue;
 
-import static org.sustain.controllers.SpatialQueryUtil.getGeometryFromGeoJson;
+import static org.sustain.db.queries.SpatialQueryUtil.getGeometryFromGeoJson;
 
 public class OsmController {
     private static final Logger log = LogManager.getLogger(OsmController.class);
