+++ src/main/java/org/sustain/db/queries/Querier.java
@@ -1,4 +1,4 @@
-package org.sustain.querier;
+package org.sustain.db.queries;
 
 import com.mongodb.client.AggregateIterable;
 import com.mongodb.client.MongoCollection;
